+++ Microsoft.DotNet.Interactive/Events/KernelEventExtensions.cs
@@ -1,0 +1,28 @@
+﻿// Copyright (c) .NET Foundation and contributors. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+using System;
+using Microsoft.DotNet.Interactive.Commands;
+
+namespace Microsoft.DotNet.Interactive.Events
+{
+    public static class KernelEventExtensions
+    {
+        public static IKernelCommand GetRootCommand(this IKernelEvent kernelEvent)
+        {
+            if (kernelEvent == null)
+            {
+                throw new ArgumentNullException(nameof(kernelEvent));
+            }
+
+            var root = kernelEvent.Command;
+
+            while (root is KernelCommandBase kb &&  kb.Parent != null)
+            {
+                root = kb.Parent;
+            }
+
+            return root;
+        }
+    }
+}
